[Mod] skip load to memory [Mod] usign same s3 function [Mod] using zero log [Mod] usign schema [Mod] skip column check [Mod] optional delte from s3

+++ pkg/destinations/redshift/insert.go
@@ -1,9 +1,9 @@
 package redshift
 
 import (
+	"fmt"
 	"path/filepath"
 
-	"fmt"
 	"os"
 	"strings"
 
@@ -13,45 +13,7 @@ import (
 	"github.com/scratchdata/scratchdata/pkg/storage/blobstore/s3"
 )
 
-func (s *RedshiftServer) getGolumnNames(table string) (map[string]bool, error) {
-	schema := "public"
-	if s.Schema != "" {
-		schema = s.Schema
-	}
-
-	sql := `
-		select "column" as column_name
-		from pg_table_def
-		where schemaname = $1 and tablename = $2
-	`
-
-	m := map[string]bool{}
-	rows, err := s.conn.Query(sql, schema, table)
-	if err != nil {
-		return nil, fmt.Errorf("getGolumnNames: cannot fetch column names: %w", err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-
-		name := ""
-		if err := rows.Scan(&name); err != nil {
-			return nil, fmt.Errorf("getGolumnNames: cannot scan column name: %w", err)
-		}
-		m[strings.ToLower(name)] = true
-	}
-	if len(m) == 0 {
-		return nil, fmt.Errorf("getGolumnNames: no columns found: %w", err)
-	}
-	return m, nil
-}
-
 func (s *RedshiftServer) createColumns(table string, jsonTypes map[string]string) error {
-	cols, err := s.getGolumnNames(table)
-	if err != nil {
-		return err
-	}
-
-	log.Printf("Found existing columns %v", cols)
 
 	for colName, jsonType := range jsonTypes {
 		colType := "VARCHAR"
@@ -68,14 +30,19 @@ func (s *RedshiftServer) createColumns(table string, jsonTypes map[string]string
 			colType = "VARCHAR"
 		}
 
-		if _, ok := cols[strings.ToLower(colName)]; !ok {
-			log.Printf("Column %s does not exist, creating it", colName)
-			sql := fmt.Sprintf("ALTER TABLE \"%s\" ADD COLUMN \"%s\" %s", table, colName, colType)
-			_, err = s.conn.Exec(sql)
-			if err != nil {
+		sql := fmt.Sprintf("ALTER TABLE \"%s\" ADD COLUMN \"%s\" %s", table, colName, colType)
+		_, err := s.conn.Exec(sql)
+		if err != nil {
+			if !strings.Contains(err.Error(), "already exists") {
+				log.Err(err).Msg("createColumns: cannot create column")
 				return err
 			}
+
 		}
+		log.Info().Msgf("Created column %s %s", colName, colType)
+		// if _, ok := cols[strings.ToLower(colName)]; !ok {
+
+		// }
 
 	}
 
@@ -124,7 +91,15 @@ func (s *RedshiftServer) InsertFromNDJsonFile(table string, filePath string) err
 
 	}
 
-	s3Client, err := s3.NewStorageWithCreds(s.S3AccessKeyId, s.S3SecretAccessKey, s.S3Bucket, s.S3Region)
+	params := make(map[string]any)
+
+	params["region"] = s.S3Region
+	params["access_key_id"] = s.S3AccessKeyId
+	params["secret_access_key"] = s.S3SecretAccessKey
+	params["bucket"] = s.S3Bucket
+	params["skipDefaultConfig"] = true
+
+	s3Client, err := s3.NewStorage(params)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Unable to create blobstore")
 	}
@@ -155,16 +130,20 @@ func (s *RedshiftServer) InsertFromNDJsonFile(table string, filePath string) err
 		return err
 	}
 
-	copyCommand := fmt.Sprintf("COPY %s FROM 's3://%s/%s' CREDENTIALS 'aws_access_key_id=%s;aws_secret_access_key=%s' FORMAT AS JSON 'auto'", table, s.S3Bucket, filePath, s.S3AccessKeyId, s.S3SecretAccessKey)
+	copyCommand := fmt.Sprintf("COPY %s FROM 's3://%s/%s' CREDENTIALS 'aws_access_key_id=%s;aws_secret_access_key=%s' FORMAT AS JSON 'auto'", s.Schema+"."+table, s.S3Bucket, filePath, s.S3AccessKeyId, s.S3SecretAccessKey)
 
 	_, err = s.conn.Exec(copyCommand)
 	if err != nil {
 		return err
 	}
+	if s.DeleteFromS3 {
+		log.Info().Msgf("Deleting file %s from S3", filePath)
+		err = s3Client.Delete(filePath)
 
-	err = s3Client.Delete(filePath)
-	if err != nil {
-		return err
+		if err != nil {
+			log.Err(err).Msg("Failed to delete file from S3")
+			return err
+		}
 	}
 	return nil
 }
